requirements.txt
pandas~=2.2
openpyxl~=3.1
pydantic~=2.9
pyyaml~=6.0
